backend/requirements.txt
# backend/requirements.txt
# Web framework
Flask==2.3.2
Flask-CORS==4.0.0
python-dotenv==1.0.0
gunicorn==20.1.0
requests==2.31.0

# Firebase
firebase-admin==6.1.0

# Google AI
google-generativeai==0.3.0

# Document Processing
PyPDF2==3.0.1
python-docx==0.8.11
openpyxl==3.1.2
python-multipart==0.0.6

# NLP & Data
numpy==1.24.3
pandas==2.0.2
scikit-learn==1.2.2
sentence-transformers==2.2.2
bertopic==0.15.0
transformers==4.30.2
